Move library.
